feat: Enhance registration and boutique management features
- Added password validation to the registration form to ensure a minimum length of 6 characters.
- Refactored the boutiques list page to utilize a controller for better state management.
- Updated the edition boutique page to include form validation and controller integration.
- Improved the home page layout with dynamic user greetings and enterprise selection options.
- Removed commented-out bottom navigation bar code for cleaner codebase.
- Introduced a new reservation API and models for handling reservations.
- Created abstract models for enterprise entities and reservations to streamline data handling.
- Implemented controllers for managing boutiques and surcursales with proper form handling.
- Enhanced the settings page with a controller for better state management.
- Developed a new bottom sheet for selecting enterprises with tabbed views for boutiques and succursales.

+++ lib/views/controllers/abstract/auth_view_controller.dart
@@ -1,21 +1,28 @@
-import 'package:app_couture/data/models/entreprise.dart';
+import 'package:app_couture/data/models/abstract/entite_entreprise.dart';
 import 'package:app_couture/data/models/user.dart';
 import 'package:app_couture/tools/components/cache.dart';
 import 'package:app_couture/tools/components/session_manager_view_controller.dart';
-import 'package:app_couture/views/static/auth/login_page.dart';
+import 'package:app_couture/tools/widgets/messages/c_choice_message_dialog.dart';
+import 'package:app_couture/views/static/auth/auth_home_page.dart';
 import 'package:get/get.dart';
 
 abstract class AuthViewController extends SessionManagerViewController {
-  set entreprise(Entreprise value) {
-    if (Get.isRegistered<Entreprise>()) {
-      Get.replace<Entreprise>(value);
+  set entite(EntiteEntreprise value) {
+    if (Get.isRegistered<EntiteEntreprise>()) {
+      Get.replace<EntiteEntreprise>(value);
     } else {
-      Get.put<Entreprise>(value, permanent: true);
+      Get.put<EntiteEntreprise>(value, permanent: true);
     }
     update();
   }
 
-  Entreprise get entreprise => Get.find<Entreprise>();
+  EntiteEntreprise get entite {
+    if (Get.isRegistered<EntiteEntreprise>()) {
+      return Get.find<EntiteEntreprise>();
+    } else {
+      return EntiteEntreprise();
+    }
+  }
 
   User get user {
     if (Get.isRegistered<User>()) {
@@ -26,24 +33,29 @@ abstract class AuthViewController extends SessionManagerViewController {
   }
 
   Future<void> logoutUser() async {
-    Cache.clear();
-    Get.deleteAll(force: true);
-    Get.offAll(() => const LoginPage());
+    final rep = await CChoiceMessageDialog.show(
+      message: "Voulez-vous vraiment vous déconnecter ?",
+    );
+    if (rep == true) {
+      Cache.clear();
+      Get.deleteAll(force: true);
+      Get.offAll(() => const AuthHomePage());
+    }
   }
 
-  List<Entreprise> get entreprises {
-    if (Get.isRegistered<List<Entreprise>>()) {
-      return Get.find<List<Entreprise>>();
+  List<EntiteEntreprise> get entites {
+    if (Get.isRegistered<List<EntiteEntreprise>>()) {
+      return Get.find<List<EntiteEntreprise>>();
     } else {
       return [];
     }
   }
 
-  set entreprises(List<Entreprise> value) {
-    if (Get.isRegistered<List<Entreprise>>()) {
-      Get.replace<List<Entreprise>>(value);
+  set entites(List<EntiteEntreprise> value) {
+    if (Get.isRegistered<List<EntiteEntreprise>>()) {
+      Get.replace<List<EntiteEntreprise>>(value);
     } else {
-      Get.put<List<Entreprise>>(value, permanent: true);
+      Get.put<List<EntiteEntreprise>>(value, permanent: true);
     }
     update();
   }
